Move GraphicsContext to its own module
